# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: downloads
    description: Download counter operations
  - name: anthropic
    description: Anthropic AI chat operations
  - name: vehicleJobs
    description: Vehicle Job + Module Swap workflow operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /downloads/{id}:
    get:
      operationId: getDownloadCount
      tags: [downloads]
      summary: Get global download count
      description: Returns the lifetime download count for a given asset id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadCount"
    post:
      operationId: incrementDownloadCount
      tags: [downloads]
      summary: Increment global download count
      description: Atomically increments and returns the lifetime download count for a given asset id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadCount"
  /anthropic/conversations:
    get:
      operationId: listAnthropicConversations
      tags: [anthropic]
      summary: List all conversations
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
          description: Optional scope key — when set, only conversations created with this scope are returned.
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicConversation"
    post:
      operationId: createAnthropicConversation
      tags: [anthropic]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnthropicConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversation"
  /anthropic/conversations/{id}:
    get:
      operationId: getAnthropicConversation
      tags: [anthropic]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
    delete:
      operationId: deleteAnthropicConversation
      tags: [anthropic]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
  /anthropic/conversations/{id}/messages:
    get:
      operationId: listAnthropicMessages
      tags: [anthropic]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicMessage"
    post:
      operationId: sendAnthropicMessage
      tags: [anthropic]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAnthropicMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
  /anthropic/module-assistant:
    post:
      operationId: moduleAssistantChat
      tags: [anthropic]
      summary: Stateless module assistant chat (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleAssistantChatBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
  /vehicle-jobs:
    get:
      operationId: listVehicleJobs
      tags: [vehicleJobs]
      summary: List vehicle jobs
      parameters:
        - name: vin
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of vehicle jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: "#/components/schemas/VehicleJob"
                required:
                  - jobs
    post:
      operationId: createVehicleJob
      tags: [vehicleJobs]
      summary: Create or upsert a vehicle job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleJobUpsert"
      responses:
        "200":
          description: Created/updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleJob"
  /vehicle-jobs/{id}:
    get:
      operationId: getVehicleJob
      tags: [vehicleJobs]
      summary: Get a vehicle job (with recent events)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vehicle job + events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleJobWithEvents"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
    patch:
      operationId: updateVehicleJob
      tags: [vehicleJobs]
      summary: Patch a vehicle job's mutable fields
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleJobPatch"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleJob"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
    delete:
      operationId: deleteVehicleJob
      tags: [vehicleJobs]
      summary: Delete a vehicle job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /vehicle-jobs/{id}/events:
    get:
      operationId: listVehicleJobEvents
      tags: [vehicleJobs]
      summary: List events for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/VehicleJobEvent"
                required:
                  - events
    post:
      operationId: appendVehicleJobEvent
      tags: [vehicleJobs]
      summary: Append a workflow event to a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleJobEventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleJobEvent"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DownloadCount:
      type: object
      properties:
        id:
          type: string
        count:
          type: integer
          minimum: 0
      required:
        - id
        - count
    AnthropicConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        scope:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    AnthropicMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateAnthropicConversationBody:
      type: object
      properties:
        title:
          type: string
        scope:
          type: string
          nullable: true
      required:
        - title
    SendAnthropicMessageBody:
      type: object
      properties:
        content:
          type: string
        moduleContext:
          $ref: "#/components/schemas/ModuleContext"
      required:
        - content
    AnthropicConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        scope:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    ModuleContext:
      type: object
      properties:
        modules:
          type: array
          items:
            type: string
        issues:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        hexSnippets:
          type: array
          items:
            type: string
    AnthropicError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    VehicleJob:
      type: object
      properties:
        id:
          type: string
        vin:
          type: string
        title:
          type: string
          nullable: true
        vehicle:
          type: object
          additionalProperties: true
          nullable: true
        status:
          type: string
        census:
          type: object
          additionalProperties: true
          nullable: true
        fixPlan:
          type: object
          additionalProperties: true
          nullable: true
        signOff:
          type: object
          additionalProperties: true
          nullable: true
        owner:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - vin
        - status
        - createdAt
        - updatedAt
    VehicleJobUpsert:
      type: object
      properties:
        id:
          type: string
        vin:
          type: string
        title:
          type: string
          nullable: true
        vehicle:
          type: object
          additionalProperties: true
          nullable: true
        status:
          type: string
          nullable: true
        owner:
          type: string
          nullable: true
      required:
        - id
        - vin
    VehicleJobPatch:
      type: object
      properties:
        title:
          type: string
          nullable: true
        vehicle:
          type: object
          additionalProperties: true
          nullable: true
        status:
          type: string
          nullable: true
        census:
          type: object
          additionalProperties: true
          nullable: true
        fixPlan:
          type: object
          additionalProperties: true
          nullable: true
        signOff:
          type: object
          additionalProperties: true
          nullable: true
        owner:
          type: string
          nullable: true
    VehicleJobEvent:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: string
        ts:
          type: string
          format: date-time
        kind:
          type: string
        module:
          type: string
          nullable: true
        payload:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - id
        - jobId
        - ts
        - kind
    VehicleJobEventInput:
      type: object
      properties:
        kind:
          type: string
        module:
          type: string
          nullable: true
        payload:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - kind
    VehicleJobWithEvents:
      type: object
      properties:
        id:
          type: string
        vin:
          type: string
        title:
          type: string
          nullable: true
        vehicle:
          type: object
          additionalProperties: true
          nullable: true
        status:
          type: string
        census:
          type: object
          additionalProperties: true
          nullable: true
        fixPlan:
          type: object
          additionalProperties: true
          nullable: true
        signOff:
          type: object
          additionalProperties: true
          nullable: true
        owner:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: "#/components/schemas/VehicleJobEvent"
      required:
        - id
        - vin
        - status
        - createdAt
        - updatedAt
        - events
    ModuleAssistantChatBody:
      type: object
      properties:
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string
            required:
              - role
              - content
        moduleContext:
          type: object
          properties:
            modules:
              type: array
              items:
                type: string
            issues:
              type: array
              items:
                type: string
            warnings:
              type: array
              items:
                type: string
            hexSnippets:
              type: array
              items:
                type: string
          required:
            - modules
            - issues
            - warnings
      required:
        - messages
        - moduleContext