# lib/api-zod/src/generated/types/vehicleJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleJobCensus } from "./vehicleJobCensus";
import type { VehicleJobFixPlan } from "./vehicleJobFixPlan";
import type { VehicleJobSignOff } from "./vehicleJobSignOff";
import type { VehicleJobVehicle } from "./vehicleJobVehicle";

export interface VehicleJob {
  id: string;
  vin: string;
  title?: string | null;
  vehicle?: VehicleJobVehicle;
  status: string;
  census?: VehicleJobCensus;
  fixPlan?: VehicleJobFixPlan;
  signOff?: VehicleJobSignOff;
  owner?: string | null;
  createdAt: Date;
  updatedAt: Date;
}