# lib/api-zod/src/generated/types/vehicleJobEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleJobEventPayload } from "./vehicleJobEventPayload";

export interface VehicleJobEvent {
  id: number;
  jobId: string;
  ts: Date;
  kind: string;
  module?: string | null;
  payload?: VehicleJobEventPayload;
}