# lib/api-zod/src/generated/types/vehicleJobEventInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleJobEventInputPayload } from "./vehicleJobEventInputPayload";

export interface VehicleJobEventInput {
  kind: string;
  module?: string | null;
  payload?: VehicleJobEventInputPayload;
}