# lib/api-zod/src/generated/types/vehicleJobPatch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleJobPatchCensus } from "./vehicleJobPatchCensus";
import type { VehicleJobPatchFixPlan } from "./vehicleJobPatchFixPlan";
import type { VehicleJobPatchSignOff } from "./vehicleJobPatchSignOff";
import type { VehicleJobPatchVehicle } from "./vehicleJobPatchVehicle";

export interface VehicleJobPatch {
  title?: string | null;
  vehicle?: VehicleJobPatchVehicle;
  status?: string | null;
  census?: VehicleJobPatchCensus;
  fixPlan?: VehicleJobPatchFixPlan;
  signOff?: VehicleJobPatchSignOff;
  owner?: string | null;
}