# lib/api-zod/src/generated/types/vehicleJobUpsert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleJobUpsertVehicle } from "./vehicleJobUpsertVehicle";

export interface VehicleJobUpsert {
  id: string;
  vin: string;
  title?: string | null;
  vehicle?: VehicleJobUpsertVehicle;
  status?: string | null;
  owner?: string | null;
}